add configure step to be able to talk directly to node-gyp
