+++ lib/node-pre-gyp.js
@@ -27,7 +27,8 @@ var commands = [
       'unpublish',
       'info',
       'testbinary',
-      'reveal'
+      'reveal',
+      'configure'
     ];
 var aliases = {};
